telegram-bot/requirements.txt
python-telegram-bot>=20.0
pynacl>=1.5.0
